Rename the core component
