--- src/PDGToolkitAPI/config.json
@@ -1,10 +1,0 @@
-﻿{
-    "grid" : {
-        "width" : 1280,
-        "height" : 720
-    },
-    "tiles" : {
-        "size" : 10
-    },
-    "outputRelativePath" : "dungeon.json"
-}
