+++ src/PDGToolkitCore/Application/RoomBuilder.cs
@@ -1,8 +1,8 @@
 ﻿using System;
 using System.Collections.Generic;
-using PDGToolkitAPI.Domain.Models;
+using PDGToolkitCore.Domain.Models;
 
-namespace PDGToolkitAPI.Application
+namespace PDGToolkitCore.Application
 {
     public class RoomBuilder
     {
