+++ src/PDGToolkitCore/Infrastructure/Runner.cs
@@ -1,10 +1,10 @@
 ﻿using System;
 using System.Threading.Tasks;
-using PDGToolkitAPI.API;
-using PDGToolkitAPI.API.Serialisers;
-using PDGToolkitAPI.Application;
+using PDGToolkitCore.API;
+using PDGToolkitCore.API.Serialisers;
+using PDGToolkitCore.Application;
 
-namespace PDGToolkitAPI.Infrastructure
+namespace PDGToolkitCore.Infrastructure
 {
     public sealed class Runner : IRunner
     {
