+++ tests/PDGToolkitAPI.UnitTests/TileBuilder.cs
@@ -1,5 +1,5 @@
 ﻿using System.Collections.Generic;
-using PDGToolkitAPI.Domain.Models;
+using PDGToolkitCore.Domain.Models;
 
 namespace PDGToolkitAPI.UnitTests
 {
